restructure displayio
