+++ src/clue/displayio/color_type.py
@@ -1,0 +1,7 @@
+class ColorType:
+    def __init__(self, rgb888):
+        self.rgb888 = rgb888
+        self.transparent = False
+
+    def get(self):
+        return self.rgb888
