+++ src/clue/displayio/group_item.py
@@ -1,0 +1,4 @@
+
+class GroupItem:
+    def draw(self, x, y, scale):
+        pass
