+++ src/clue/terminalio.py
@@ -1,4 +1,6 @@
 from adafruit_bitmap_font import bitmap_font  # pylint: disable=wrong-import-position
 import sys
+import os
 
-FONT = bitmap_font.load_font(f"{sys.path[0]}\\..\\ter-u12n.bdf")
+FONT = bitmap_font.load_font(os.path.join(sys.path[0], "..", "ter-u12n.bdf"))
+
